% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_neighbors}
\alias{get_neighbors}
\title{Build a data.frame of neighbor points}
\usage{
get_neighbors(lon, lat, win_size, pix_res)
}
\arguments{
\item{lon}{A length-one numeric. A longitude.}

\item{lat}{A length-one numeric. A latitude.}

\item{win_size}{A length-one integer. A window size.}

\item{pix_res}{A length-two list of numeric. A pixel's spatial resolution}
}
\value{
A data.frame
}
\description{
Build a data.frame of neighbor points
}
