% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{process_ts}
\alias{process_ts}
\title{Format the time series comming from WTSS}
\usage{
process_ts(ts_ls, s_points)
}
\arguments{
\item{ts_ls}{A list of time series (from wtss::listTimeSeries)}
}
\value{
A data.frame.
}
\description{
Format the time series comming from WTSS
}
