% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ts2raster}
\alias{ts2raster}
\title{Convert time series to RasterBrick}
\usage{
ts2raster(dataInput)
}
\arguments{
\item{dataInput}{A data.frame of time series (from process_ts)}
}
\value{
A list of RasterBrick
}
\description{
Convert time series to RasterBrick
}
